requirements.txt
bandit==1.7.5
blinker==1.6.2
boto3==1.28.25
botocore==1.31.25
cfgv==3.3.1
click==8.1.6
distlib==0.3.7
duckdb==0.8.0
filelock==3.12.2
Flask==2.3.2
gitdb==4.0.10
GitPython==3.1.32
gunicorn==21.2.0
identify==2.5.26
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
nodeenv==1.8.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pbr==5.11.1
platformdirs==3.10.0
polars==0.18.13
pre-commit==3.3.3
pyarrow==12.0.1
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
rich==13.5.2
s3transfer==0.6.1
six==1.16.0
smmap==5.0.0
stevedore==5.1.0
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.24.3
Werkzeug==2.3.6
